major update
Simplified parts of the configuration file, added Saved & Loaded events
Wrapped JSON save/load in try-catch, clarified file access permissions
Added a log message when a target is successfully resolved
Replaced ITargetable with IAudioControllable that actually has members
Added AudioMeterInformation properties to audio Devices & Sessions
Implemented an update timer in AudioDeviceCollection for peak meters
Replaced the config entry for the current Target with a TargetInfo struct
Added tracking of Session Instance Identifiers to improve target logic
Removed 'Endpoint' from audio device property names so they're consistent
Added InputSimulatorEx nuget package to VolumeControl.Hotkeys
Added JSON serialization support for ObservableImmutableList
Replaced some config entry types with ObservableImmutableList
Added a Hyperlink extension property that opens the link in the default browser
Fixed #40 & replaced NotifyIcon implementation with a more sane one
VolumeControlNotifyIcon will automatically update when the language changes
Clarified some of the addon loading log messages
Reworked the version getter to allow upgrade detection so translations can be updated
Reworked the RunAtStartup feature to indicate when the registry key is set for another Instance
Added full multi-instance support; see the `AllowMultipleDistinctInstances` setting
Added a better focus indicator to some GUI elements
Added a basic about tab (#35)
Added translations for all new GUI elements that have text

+++ VolumeControl.Hotkeys/Enum/EVirtualKeyCode.cs
@@ -6,7 +6,7 @@
     /// This includes every labelled virtual key used by windows.<br/>
     /// <see href="https://docs.microsoft.com/en-us/windows/win32/inputdev/virtual-key-codes"/>
     /// </summary>
-    public enum EVirtualKeyCode : byte
+    public enum EVirtualKeyCode : short
     {
         /// <summary>Null</summary>
         None = 0x00,
